centralize cfssl loglevel setup
